refactor(notebook): lift outline interaction into shared hooks
Extract useActiveOutlineItemId and useOutlineSelection from
apps/notebook/src/App.tsx into src/components/notebook/outline-interaction.ts.
These hooks manage scroll-synced active item tracking and
outline-to-focused-cell selection coupling, used by both desktop and cloud.

Also extract getOutlineStatusLabel as useOutlineStatusLabel(), built on the
shared useNotebookQueueProjection, into outline-status-label.ts.

Desktop App.tsx switches from ~160 lines of local implementation to
importing the shared hooks. Behavior-preserving: the extraction is verbatim,
and desktop tests pass with the new imports.

Tests in outline-interaction.test.ts cover selection-clearing on focus
change, focus coupling when setFocusedCellId is provided, and graceful
degradation when setFocusedCellId is undefined.
